+++ srcs/init/init_mlx.c
@@ -1,0 +1,49 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init_mlx.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: myokono <myokono@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 10:38:23 by myokono           #+#    #+#             */
+/*   Updated: 2025/04/29 16:39:32 by myokono          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/cub3d.h"
+
+void	init_game(t_game *game)
+{
+	ft_memset(game, 0, sizeof(t_game));
+	game->floor_color = -1;
+	game->ceiling_color = -1;
+}
+
+int init_mlx(t_game *game)
+{
+	game->mlx = mlx_init();
+	if (!game->mlx)
+		return (error_msg(ERR_MEMORY));
+	game->win = mlx_new_window(game->mlx, WINDOW_WIDTH, WINDOW_HEIGHT, TITLE);
+	if (!game->win)
+	{
+		mlx_destroy_display(game->mlx);
+		free(game->mlx);
+		game->mlx = NULL;
+		return (error_msg(ERR_MEMORY));
+	}
+	game->img.img = mlx_new_image(game->mlx, WINDOW_WIDTH, WINDOW_HEIGHT);
+	if (!game->img.img)
+	{
+		mlx_destroy_window(game->mlx, game->win);
+		mlx_destroy_display(game->mlx);
+		free(game->mlx);
+		game->mlx = NULL;
+		game->win = NULL;
+		return (error_msg(ERR_MEMORY));
+	}
+	game->img.addr = mlx_get_data_addr(game->img.img, &game->img.bits_per_pixel,
+									  &game->img.line_length, &game->img.endian);
+	
+	return (1);
+}
